Use webdav.FileSystem interface instead of direct file IO.
This allows using an in-memory store as a default (when a permanent
on-disk snippet store location is not provided), which is a better
default value when someone wants just to try out the binary without it
writing to disk (even if a temporary folder).

It's a marginal improvement in utility and mostly an improvement in
code style. In theory, this allows providing a different filesystem-like
implementation for a snippet store.