refactor top-level code organization
Make changes to improve readability and modernize top-level
code organization.

This change eleminates global variables to help readability.
There is now a snippet store abstraction, and an HTTP server
abstraction. Both are backed by struct types with methods.
As a result, it's easier to see all of the dependencies for
each component of code.